requirements.txt
# Core
crewai>=0.28.0
crewai-tools>=0.2.0

# LLM
openai>=1.12.0
anthropic>=0.18.0

# Config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Letta AI - Solo client HTTP (no dipendenze pesanti)
requests>=2.31.0
httpx>=0.24.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# Dev
pytest>=7.4.0
black>=23.12.0
